Refactor/safe start

+++ src/core/modules/project_management/api/desktop/portfolio/commands/create_scenario.py
@@ -1,11 +1,12 @@
 from __future__ import annotations
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass(frozen=True)
 class PortfolioScenarioCreateCommand:
     name: str
-    budget_limit: float | None = None
+    budget_limit: Decimal | None = None
     capacity_limit_percent: float | None = None
     project_ids: tuple[str, ...] = ()
     intake_item_ids: tuple[str, ...] = ()
